hw6/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pg8000>=1.30.0
google-cloud-storage>=2.10.0
